[Testing] Allow ingest view fixture data to be transformed for parser test. (Recidiviz/recidiviz-data#30806)
I was running into an issue where fixture data was being (very
helpfully) cast into types I did not want by the csv.DictReader. This
change allows us to specify an ingest_view column name and callable to
ensure it is the correct type. I note an issue to get this information
from YAMLs in the future.

GitOrigin-RevId: 784a7685682a842aa141a6966848863e0d289647