[Insights] Change insights_supervision_officers_archive file source (Recidiviz/recidiviz-data#30814)
## Description of the change

Updates the filename in the `source_uri` for the
`export_archives.insights_supervision_officers_archive` archive. It was
pointing at the wrong file so the archive has been empty for a while.

I'm a little concerned that the `terraform.lock.hcl` didn't change
during the commit hook. @danawillow Is this concerning for you?

I am definitely not well versed in Terraform. I did not run `init` or
`plan`...do I need to for this simple change?

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4ed603e6a7e297fbe1f4a76bf17f97d59983dc25