[US_MO] MOSOP list updates (Recidiviz/recidiviz-data#30693)
## Description of the change

Updates the MOSOP list with a slight tweak to the output and a fix for a
previous bug:

- No longer shows prioritized dates from people serving life sentences
- People who have a reconsideration hearing date and are past their
ME/MPT dates should have their prioritized date set to either the RCH
date or the CR/MAX date, depending on how distant the RCH date is. This
part of the logic was incorrectly overriding the prioritized date with
the CR/MAX date for people who don't have an RCH date, even though the
board-determined release date should take priority in these cases. This
has been fixed, so the override no longer applies to people without a
RCH date.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: ccbdb602ec3e0288e38ae0a7dc7443bca293eb8f