[Insights] Add new officer sessions view to handle transitional specialist issue (Recidiviz/recidiviz-data#30432)
## Description of the change
**Goal deadline to have this validated and merged week of 6/17**

This PR creates a new supervision officer sessions view, that identifies
for officers with "transitional" caseload types (ie p&p specialists,
dummy parole commission officers, etc) the most recent previous non null
"active" officer within that legal authority session. **The output of
this new view includes a new span for the non transitional officer that
encompasses the period of time where the client was on a transitional
caseload, and retains the supervision officer session for the
transitional officer.**

The goal of this view is to assign person events, such as incarcerations
and absconsions to the officer that was supervising the client BEFORE
they were transferred to the transitional officer, while also providing
the flexibility to attribute positive events such as early discharges to
the transitional officer.

specifically:
- I create `supervision_officer_transitional_caseload_type_sessions`
which references a new view
`supervision_staff_primary_specialized_caseload_type_sessions`
- Updates `metric_unit_of_analysis_type` to reference this view

**Validation** The only rows that changed from
`supervision_officer_sessions` in Recidiviz/recidiviz-data#30567 and
`supervision_officer_transitional_caseload_type_sessions` are for
`US_IX`. We see a <1% difference.

Even so, 50% of the
[changes](https://docs.google.com/spreadsheets/d/12D0FF3JvUdyy5B_lHk8Z5ZPPAm8ndKfS8vwAyr6AsX0/edit?gid=321741388#gid=321741388)
were due to the fact that we only started ingesting specialized caseload
type on `2023-02-21` and so we see periods for `TRANSITIONAL` officers
split up on this date.

![image](https://github.com/Recidiviz/recidiviz-data/assets/93396549/b8acf617-3690-4334-9e94-f3b9bdbfa2f2)

**note to sam: taking the last non null officer resulted in 231 fewer
overall changes, and fewer new non null periods. Other changes remained
basically the same**

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#29249

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d222ac29b6b47a3e22197c725df1b2bfc3b12e74