[US_MI] Revise absconsion mapping to ignore "Probation Absconder Warrant" and "Parole Absconder Warrant" schedules (Recidiviz/recidiviz-data#30621)
## Description of the change

Previously, we would map absconsion type and absconsion level to
`ABSCONSION` if we saw "Probation Absconder Warrant" or "Parole
Absconder Warrant" in the `supervision_level_id` field we create.
However, we've been told by TTs that those are not reliable indicators
for absconsion, and instead we should just look at the modifier field.
This PR revises the absconsion type and absconsion level mappings to
ignore "Probation Absconder Warrant" and "Parole Absconder Warrant".

In a later PR, we will also revise the absconsion mappings to ignore
movement reasons:
https://github.com/Recidiviz/recidiviz-data/issues/30848

Testing
- [x] Sandbox
  - [x] Successful
  - [x] re-validate differences

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 584d0d22e1f529cdafc24453702c0d46a3009d16