[US_PA] Ingest task deadlines for parole sentences (Recidiviz/recidiviz-data#30764)
## Description of the change

Ingest `DISCHARGE_FROM_SUPERVISION` task deadlines for parole sentences.

The supervision discharge is stored either in the `RelMaxDate` field on
the `dbo_ReleaseInfo` table when the client is still actively on parole,
or on the `HReMaxDate` field in `dbo_Hist_Release` if the client has
completed parole and their record has been archived. Therefore, when we
have record of both dates, we'll defer to what should have been the most
final date in the `dbo_Hist_Release` table.

Should fix most of Recidiviz/recidiviz-data#28858

Testing:
- [x] Run sandbox
- [ ] After this merges in, check that the number of errant expiration
dates goes down

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 9f0de6299a9684806c6e40702ab492ec4ddb9473