[Justice Counts] Add invariants for when a MetricInterface's aggregated_dimension field is empty. (Recidiviz/recidiviz-data#30822)
## Description of the change

Add invariants for when a MetricInterface's aggregated_dimension field
is empty.

This really only occurs after the creation of new agencies.

This copies the existing behavior for when there are no agency
datapoints in the table (after the creation of a new agency) - the
equivalent existing logic is in datapoints_for_metric.py on line 279.

## Testing

I tested this locally and on a playtest deployment! This fixes all of
the issues we were running into during Thursday's playtest session!

GitOrigin-RevId: f022f5de21d1a7e16cc239e9224e4894ab914551