[BigQuery] Update ingest view schemas after raw data table schemas (Recidiviz/recidiviz-data#30962)
## Description of the change
Today's production deploy failed as it introduced an ingest view change
that relied on a column that did not yet exist in raw data.

The sequencing of `update_big_query_table_schemas` previously built
ingest view schema prior to performing any raw data table updates, which
resulted in the following error:

```
>  google.api_core.exceptions.BadRequest: 400 Unrecognized name: CURRENT_NEEDS at [19:166]; reason: invalidQuery, location: query, message: Unrecognized name: CURRENT_NEEDS at [19:166]
```

This changes the sequencing so that ingest view output tables are
updated separately.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 2d9c181adf0343c9e8247991bfa6196612b48a6d