Refactor time period aggregation for year/quarter metrics, and remove weekly metrics (Recidiviz/recidiviz-data#30949)
## Description of the change

This PR makes two changes:
1) Previously we calculated MONTH metrics and reaggregated them to the
QUARTER and YEAR level -- now we are just calculating quarterly and
yearly metrics directly for period span metrics. This fixes a bug where
we were incorrectly handling/aggregating year metrics for units of
analysis that did not have full coverage of months in the analysis
period.
2) We remove WEEK metrics since these don't have any product use cases
and can be recreated in the custom looker metric explores

Impacted insights officers who would have previously not been surfaced
as outliers but now are:
https://docs.google.com/spreadsheets/d/1CnEmLnmVw0Nxy5x__dpQ6bDgpkrssafhHx9djrxYlvY/edit?gid=0#gid=0

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#30948
Closes Recidiviz/recidiviz-data#30717

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 92c78e9752d7aab54898db47f971ee18d30b9810