[Supervisor homepage] Add querier method to pull in excluded officers for a supervisor (Recidiviz/recidiviz-data#30831)
There are currently two types of officer exclusions used in the insights views - supervision_officer_metric_exclusions and supervision_staff_exclusions. these exclusions can be specified in the state configs, and are applied to the supervision_officer_metrics view and the supervision_officers view respectively. A P0 for the july launch of the new homepage is to pull in the officers excluded from the ``supervision_officer_metric` view and display them as "excluded" in the FE.

Since the SOM view is based on the SO view, but with the supervision_officer_metric_exclusions applied, the difference in officers between those two views, should be the excluded officers that we're looking for.

GitOrigin-RevId: 1868d0916a105f23f0bf2b84cd8f6ca110ae6e56