Allow viewing multiple task types at once in workflows impact explore (Recidiviz/recidiviz-data#30743)
## Description of the change

This change enables users to aggregate impact metrics across multiple
task types at once in the workflows impact metrics explore. It
accomplishes this by converting the `task_type` parameter (which only
allows one selection at a time) into a filter field associated with a
`task_type` dimension that pulls from a json-extracted task type field
in the metric subqueries.

This meant adjusting the existing lookml functions to accept a new
parameter called "json_field_filters", which is a list of string column
names that are assumed to be able to be sourced from json blobs (either
event_attributes or span_attributes). Each field in this inputted list
is json-extracted within the metric subqueries so that it can be
referenced via its own dimension and associated filter.

See [design
doc](https://www.notion.so/recidiviz/Design-Doc-Allowing-cross-opportunity-workflows-impact-aggregations-in-new-looker-545092dc7e01413eb667e00fa3c6dad1)
for more!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

---------

Co-authored-by: mayukas <mayuka@stanford.edu>
GitOrigin-RevId: 826ea93cb602fb5e6f44f712854db43ba8bd7bf7