[US_ND] Update raw data update cadence (Recidiviz/recidiviz-data#31047)
## Description of the change

The DOCR business analyst team has suggested that the
`recidiviz_elite_assessment_results` table is no longer in use. This PR
changes the `update_cadence` on the raw data config for that file so
that it does not get flagged as stale when it is not sent daily. We are
not currently using this table in any ingest or analyst views, so it is
not an urgent problem that it is no longer in use. I have it on the
agenda to discuss where the data it contained will be stored in the
future with the BA team.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Related to Recidiviz/recidiviz-data#29979

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 397d557e41c465aa8c6245085c275d10fe9fb4eb